+++ backend/src/auth/auth.service.ts
@@ -94,17 +94,24 @@ export class AuthService {
       throw new HttpException(`User with this email not fount`, HttpStatus.NOT_FOUND);
     }
     const code = uuid.v4();
+    console.log(user.verificationCode);
     user.verificationCode = code;
+    console.log(user.verificationCode);
+
     await user.save();
     await this.mailService.sendCode(user.email, code, user.fullname);
   }
 
   async verifyCode(email: string, code: string) {
     const user = await this.userService.getByEmail(email);
+    console.log(user.verificationCode);
     if (!user) {
       throw new HttpException(`User with this email not fount`, HttpStatus.NOT_FOUND);
     }
-    return user.verificationCode === code;
+    if (user.verificationCode === code) {
+      return true;
+    }
+    throw new HttpException(`Incorrect code`, HttpStatus.BAD_REQUEST);
   }
 
   async changePassword(email: string, code: string, password: string) {