+++ frontend/src/components/Header/Header.tsx
@@ -20,7 +20,7 @@ export const Header: React.FC<IHeader> = ({ isAuth, username, navbar, setNavbar 
 
 	return (
 		<Box sx={{ flexGrow: 1 }}>
-			<AppBar position="static">
+			<AppBar position="static" color="primary">
 				<Toolbar>
 					<IconButton
 						onClick={() => setNavbar(!navbar)}